// packages/shared/src/generated/ObjectActivation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CheckpointBlobManifest } from "./CheckpointBlobManifest";
import type { DocumentDescriptor } from "./DocumentDescriptor";
import type { EncryptedCheckpoint } from "./EncryptedCheckpoint";
import type { PolicyEnvelope } from "./PolicyEnvelope";

/**
 * Atomic publication of a writer-created collaborative object.
 */
export type ObjectActivation = { version: 1 | 2, activationId: string, workspaceId: string, policyRevision: string, coveredSequence: string, capabilityDigest: string, deviceId: string, document: DocumentDescriptor, envelopes: Array<PolicyEnvelope>, checkpoint: EncryptedCheckpoint, blobs?: Array<CheckpointBlobManifest>, signature: string, };